test(tron): the guarded sends live where a test can reach them in CI
Both new suites imported the CLI script they cover, whose chain reaches
`demoScriptHelpers.ts` and from there `typechain` — a generated directory the
ts-tests job does not build. So `run-ts-tests` failed with `Cannot find module
'../../../typechain'` while the same suites passed locally, where a stray
`typechain` symlink in the worktree was standing in for it.

`sendTransferOwnership` and `sendGuardedFacetRegistration` move into
`send-transfer-ownership.ts` and `send-guarded-facet-registration.ts`, which
import only the pre-flight and the constants they need. The two suites are
renamed to match, so each once again sits beside the module it covers.

Verified by hiding the symlink and re-running: 2445 pass / 0 fail with and
without `typechain` present, against the 2434 / 2 that CI reported.
